clean Analitic
