+++ app.js
@@ -1,7 +1,5 @@
 const express = require('express');
-const fetch = require('node-fetch');
-// import fetch from 'node-fetch';
-// const expressGa = require('express-ga-middleware');
+const analytic = require('./middlewares/analytics');
 const app = express();
 var favicon = require('serve-favicon');
 app.use(favicon(__dirname + '/public/favicon.ico'));
@@ -9,34 +7,14 @@ require('dotenv').config();
 const {
     PORT = 3110
 } = process.env;
-const { GA_TRACKING_ID } = process.env;
 const morgan = require('morgan');
 app.use(morgan('dev'));
 app.use(express.json());
 app.enable('trust proxy');
 // app.use(expressGa('G-MRYF58C8K7'));
-app.use("/image", express.static("public/imagekit"));
-app.use("/img/icon", express.static("public/img/icon"));
-app.use("/img", express.static("public/img/cdn"));
-app.use("/imgs", async (req, res, next) => {
-    console.log('image');
-    // console.log(req.hostname);
-    console.log(req.ip);
-    console.log(req.path);
-    console.log(req.hostname);
-    console.log(req.headers['user-agent']);
-    try {
-        console.log(await trackEvent(req.ip, 'Category', 'Action', 'Event Label', '55', req.path));
-
-        // next();
-        console.log('success');
-    } catch (err) {
-        console.log('error');
-        // console.log(err);
-    }
-
-    next();
-}, express.static("public/img/cdn"));
+app.use("/image", analytic.analytics, express.static("public/imagekit"));
+app.use("/img/icon", analytic.analytics, express.static("public/img/icon"));
+app.use("/img", analytic.analytics, express.static("public/img/cdn"));
 
 // welcome
 app.get('/', (req, res) => {
@@ -62,35 +40,3 @@ app.listen(PORT, () => {
     console.log('listening on port', PORT);
     console.log(process.env.baseURL);
 });
-const trackEvent = (cid, category, action, label, value, document_path) => {
-    const data = {
-        // API Version.
-        v: '1',
-        // Tracking ID / Property ID.
-        tid: GA_TRACKING_ID,
-        // Anonymous Client Identifier. Ideally, this should be a UUID that
-        // is associated with particular user, device, or browser instance.
-        // cid: '555',
-        cid: cid,
-        // Event hit type.
-        t: 'pageview',
-        // Event category.
-        ec: category,
-        // Event action.
-        ea: action,
-        // Event label.
-        el: label,
-        // Event value.
-        ev: value,
-        // Document path.
-        dp: document_path
-    };
-
-    return fetch('http://www.google-analytics.com/collect', {
-        params: data,
-        headers: {
-            'user-agent':
-                'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36',
-        },
-    });
-};
